Struct turbo_tasks_env::DotenvProcessEnvVc

pub struct DotenvProcessEnvVc { /* private fields */ }

Vc for [DotenvProcessEnv]

A reference to a value created by a turbo-tasks function. The type can either point to a cell in a [turbo_tasks::Task] or to the output of a [turbo_tasks::Task], which then transitively points to a cell again, or to an fatal execution error.

Self::resolve().await? can be used to resolve it until it points to a cell. This is useful when storing the reference somewhere or when comparing it with other references.

A reference is equal to another reference when it points to the same thing. No resolving is applied on comparison.

Implementations

impl DotenvProcessEnvVc

pub fn as_value_debug(self) -> ValueDebugVc

impl DotenvProcessEnvVc

pub async fn resolve(self) -> Result<Self>

see [turbo_tasks::RawVc::resolve]

pub async fn resolve_strongly_consistent(self) -> Result<Self>

see [turbo_tasks::RawVc::resolve_strongly_consistent]

pub async fn resolve_from( super_trait_vc: impl Into<RawVc> ) -> Result<Option<Self>, ResolveTypeError>

pub fn strongly_consistent(self) -> ReadRawVcFuture<ValueCast<DotenvProcessEnv>>

The invalidation of tasks due to changes is eventually consistent by default. Tasks will execute as early as any of their children has changed, even while other children or grandchildren are still computing (and may or may not result in a future invalidation). Tasks may execute multiple times until the graph reaches the end state. Partial applied changes might be visible to the user. But changes are available as fast as possible and won’t be blocked by some slower parts of the graph (e. g. recomputation of blurred images, linting, etc).

When you read a task with .strongly_consistent() it will make that one read operation strongly consistent. That means it will only return a result when all children and grandchildren in that graph have been settled. This means your current task will recompute less often, but it might also need to wait for slower operations in the graph and can’t continue with partial applied changes.

Reading strongly consistent is also far more expensive compared to normal reading, so it should be used with care.

impl DotenvProcessEnvVc

pub fn new(prior: Option<ProcessEnvVc>, path: FileSystemPathVc) -> Self

impl DotenvProcessEnvVc

pub fn read_prior(self) -> EnvMapVc

impl DotenvProcessEnvVc

pub fn read_all_with_prior(self, prior: EnvMapVc) -> EnvMapVc

impl DotenvProcessEnvVc

pub fn as_process_env(self) -> ProcessEnvVc